fix(relearn): render the Relearn tab in place instead of a separate page
Clicking the Relearn tab navigated to a bare /quiz/relearn page with no tab
row — so it wasn't really a tab, and the tabbed layout was gone. Make it a true
in-place tab like Vocabulary/Etymology:

- Extract the start screen into a RelearnStart component; the hub renders it
  inline under the tab row when the Relearn tab is active (no navigation).
- Delete the standalone /quiz/relearn page.
- session (empty queue) and complete ("Relearn again") return via
  /quiz?tab=relearn; the hub reads that param to reopen the Relearn tab.

Tests moved to RelearnStart.test.tsx; redirect assertions and the e2e step
updated; COVERAGE.md drops the removed route.
