+++ frontend/e2e/steps/relearn.steps.ts
@@ -3,11 +3,11 @@ import { createBdd } from "playwright-bdd";
 
 const { When, Then } = createBdd();
 
-// covers route: /quiz/relearn
+// The Relearn tab switches content in place within the Quiz hub (no navigation).
 When("I open the Relearn Quiz", async ({ page }) => {
   await page.goto("/quiz");
   await page.getByText("Relearn", { exact: true }).click();
-  await expect(page).toHaveURL(/\/quiz\/relearn$/);
+  await expect(page.getByText("Look back over the last:")).toBeVisible();
 });
 
 Then("I see words to relearn", async ({ page }) => {
