+++ frontend/src/app/quiz/page.tsx
@@ -20,6 +20,7 @@ import {
   type NotebookSummary,
 } from "@/lib/client";
 import { useQuizStore, type QuizType } from "@/store/quizStore";
+import RelearnStart from "@/components/RelearnStart";
 
 type Tab = "vocabulary" | "etymology" | "relearn";
 type VocabMode = "standard" | "reverse" | "freeform";
@@ -77,6 +78,14 @@ export default function QuizHubPage() {
   // driven refetches keep the current list visible and swap counts in
   // place when the response arrives — toggling the switch shouldn't feel
   // like a page reload. initialLoadRef gates that distinction.
+  // Open the Relearn tab when arrived at via /quiz?tab=relearn (from the
+  // relearn complete screen's "Relearn again", or a session that ended empty).
+  useEffect(() => {
+    if (new URLSearchParams(window.location.search).get("tab") === "relearn") {
+      setTab("relearn");
+    }
+  }, []);
+
   const initialLoadRef = useRef(true);
   useEffect(() => {
     if (initialLoadRef.current) {
@@ -265,12 +274,6 @@ export default function QuizHubPage() {
     }, 0);
 
   const handleTabChange = (newTab: Tab) => {
-    // Relearn is a separate, cross-quiz-type flow with its own start screen, so
-    // its tab navigates there rather than swapping mode cards in place.
-    if (newTab === "relearn") {
-      router.push("/quiz/relearn");
-      return;
-    }
     setTab(newTab);
     setSelectedIds(new Set());
     setSectionSelections(new Map());
@@ -458,7 +461,11 @@ export default function QuizHubPage() {
         ))}
       </Box>
 
-      {/* Mode cards */}
+      {/* Content: the Relearn tab renders its start screen inline (so it stays
+          within the tabbed hub); the other tabs show mode cards + notebooks. */}
+      {tab === "relearn" ? (
+        <RelearnStart />
+      ) : (
       <Box p={4} display="flex" flexDirection="column" gap={3}>
         {modes.map((mode) => {
           const isSelected = selectedMode === mode.key;
@@ -675,6 +682,7 @@ export default function QuizHubPage() {
           <Text color="red.500" mt={2}>{error}</Text>
         )}
       </Box>
+      )}
     </Box>
   );
 }
