+++ frontend/src/components/RelearnStart.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { render, screen, fireEvent, waitFor } from "@testing-library/react";
 import { ChakraProvider, defaultSystem } from "@chakra-ui/react";
-import RelearnStartPage from "./page";
+import RelearnStart from "./RelearnStart";
 import { useRelearnStore } from "@/store/relearnStore";
 import type { RelearnCard } from "@/lib/client";
 
@@ -16,17 +16,17 @@ vi.mock("@/lib/client", () => ({
 const pushMock = vi.fn();
 vi.mock("next/navigation", () => ({ useRouter: () => ({ push: pushMock }) }));
 
-function renderPage() {
+function renderStart() {
   return render(
     <ChakraProvider value={defaultSystem}>
-      <RelearnStartPage />
+      <RelearnStart />
     </ChakraProvider>,
   );
 }
 
 const card = (entry: string): RelearnCard => ({ entry, noteId: BigInt(1) }) as RelearnCard;
 
-describe("RelearnStartPage", () => {
+describe("RelearnStart", () => {
   beforeEach(() => {
     startRelearnQuiz.mockReset();
     pushMock.mockReset();
@@ -35,21 +35,21 @@ describe("RelearnStartPage", () => {
 
   it("loads the pool and shows the word count", async () => {
     startRelearnQuiz.mockResolvedValue({ cards: [card("a"), card("b")] });
-    renderPage();
+    renderStart();
     expect(await screen.findByText("2 words to relearn")).toBeInTheDocument();
     expect(startRelearnQuiz).toHaveBeenCalledWith({ windowHours: 24 });
   });
 
   it("shows the empty state when nothing is pooled", async () => {
     startRelearnQuiz.mockResolvedValue({ cards: [] });
-    renderPage();
+    renderStart();
     expect(await screen.findByText("Nothing to relearn.")).toBeInTheDocument();
     expect(screen.getByRole("button", { name: "Start" })).toBeDisabled();
   });
 
   it("reloads with a different window when a preset is chosen", async () => {
     startRelearnQuiz.mockResolvedValue({ cards: [card("a")] });
-    renderPage();
+    renderStart();
     await screen.findByText("1 word to relearn");
     fireEvent.click(screen.getByText(/6 hours/));
     await waitFor(() => expect(startRelearnQuiz).toHaveBeenLastCalledWith({ windowHours: 6 }));
@@ -57,7 +57,7 @@ describe("RelearnStartPage", () => {
 
   it("seeds the queue and navigates on Start", async () => {
     startRelearnQuiz.mockResolvedValue({ cards: [card("a"), card("b")] });
-    renderPage();
+    renderStart();
     await screen.findByText("2 words to relearn");
     fireEvent.click(screen.getByRole("button", { name: "Start" }));
     expect(useRelearnStore.getState().queue.map((c) => c.entry)).toEqual(["a", "b"]);
@@ -66,7 +66,7 @@ describe("RelearnStartPage", () => {
 
   it("shows an error when loading fails", async () => {
     startRelearnQuiz.mockRejectedValue(new Error("boom"));
-    renderPage();
+    renderStart();
     expect(await screen.findByRole("alert")).toHaveTextContent("Failed to load");
   });
 });
